Safari-609.1.11.1/LayoutTests/fast/canvas/canvas-transparency-and-composite-expected.txt

Series of tests to ensure correct behaviour of composite on a fully transparent color.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS imgdata[4] is 0
PASS imgdata[5] is 128
PASS imgdata[6] is 0
PASS successfullyParsed is true

TEST COMPLETE
